+++ components/button.tsx
@@ -1,4 +1,4 @@
-import { cls } from "../libs/utils";
+import { cls } from '../libs/client/utils';
 
 interface ButtonProps {
   large?: boolean;
@@ -6,19 +6,14 @@ interface ButtonProps {
   [key: string]: any;
 }
 
-export default function Button({
-  large = false,
-  onClick,
-  text,
-  ...attrs
-}: ButtonProps) {
+export default function Button({ large = false, onClick, text, ...attrs }: ButtonProps) {
   return (
     <button
       {...attrs}
       onClick={onClick}
       className={cls(
-        "w-full bg-orange-500 hover:bg-orange-600 text-white  px-4 border border-transparent rounded-md shadow-sm font-medium focus:ring-2 focus:ring-offset-2 focus:ring-orange-500 focus:outline-none",
-        large ? "py-3 text-base" : "py-2 text-sm "
+        'w-full bg-orange-500 hover:bg-orange-600 text-white  px-4 border border-transparent rounded-md shadow-sm font-medium focus:ring-2 focus:ring-offset-2 focus:ring-orange-500 focus:outline-none',
+        large ? 'py-3 text-base' : 'py-2 text-sm ',
       )}
     >
       {text}
