+++ components/layout.tsx
@@ -1,6 +1,6 @@
-import Link from "next/link";
-import { useRouter } from "next/router";
-import { cls } from "../libs/utils";
+import Link from 'next/link';
+import { useRouter } from 'next/router';
+import { cls } from '../libs/client/utils';
 
 interface LayoutProps {
   title?: string;
@@ -9,12 +9,7 @@ interface LayoutProps {
   children: React.ReactNode;
 }
 
-export default function Layout({
-  title,
-  canGoBack,
-  hasTabBar,
-  children,
-}: LayoutProps) {
+export default function Layout({ title, canGoBack, hasTabBar, children }: LayoutProps) {
   const router = useRouter();
   const onClick = () => {
     router.back();
@@ -24,8 +19,8 @@ export default function Layout({
     <div>
       <header
         className={cls(
-          !canGoBack ? "justify-center" : "",
-          "bg-white w-full max-w-xl text-lg px-10 font-medium py-3 fixed text-gray-800 border-b top-0 flex items-center"
+          !canGoBack ? 'justify-center' : '',
+          'bg-white w-full max-w-xl text-lg px-10 font-medium py-3 fixed text-gray-800 border-b top-0 flex items-center',
         )}
       >
         {canGoBack ? (
@@ -37,27 +32,20 @@ export default function Layout({
               viewBox="0 0 24 24"
               xmlns="http://www.w3.org/2000/svg"
             >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M15 19l-7-7 7-7"
-              />
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" />
             </svg>
           </button>
         ) : null}
         {title ? <span>{title}</span> : null}
       </header>
-      <div className={cls("pt-12", hasTabBar ? "pb-24" : "")}>{children}</div>
+      <div className={cls('pt-12', hasTabBar ? 'pb-24' : '')}>{children}</div>
       {hasTabBar ? (
         <nav className="bg-white w-full max-w-xl text-gray-700 border-t fixed bottom-0 px-10 pb-5 pt-3 flex justify-between text-xs">
           <Link href="/">
             <a
               className={cls(
-                "flex flex-col items-center space-y-2 ",
-                router.pathname === "/"
-                  ? "text-orange-500"
-                  : "hover:text-gray-500 transition-colors"
+                'flex flex-col items-center space-y-2 ',
+                router.pathname === '/' ? 'text-orange-500' : 'hover:text-gray-500 transition-colors',
               )}
             >
               <svg
@@ -80,10 +68,8 @@ export default function Layout({
           <Link href="/community">
             <a
               className={cls(
-                "flex flex-col items-center space-y-2 ",
-                router.pathname === "/community"
-                  ? "text-orange-500"
-                  : "hover:text-gray-500 transition-colors"
+                'flex flex-col items-center space-y-2 ',
+                router.pathname === '/community' ? 'text-orange-500' : 'hover:text-gray-500 transition-colors',
               )}
             >
               <svg
@@ -106,10 +92,8 @@ export default function Layout({
           <Link href="/chats">
             <a
               className={cls(
-                "flex flex-col items-center space-y-2 ",
-                router.pathname === "/chats"
-                  ? "text-orange-500"
-                  : "hover:text-gray-500 transition-colors"
+                'flex flex-col items-center space-y-2 ',
+                router.pathname === '/chats' ? 'text-orange-500' : 'hover:text-gray-500 transition-colors',
               )}
             >
               <svg
@@ -132,10 +116,8 @@ export default function Layout({
           <Link href="/live">
             <a
               className={cls(
-                "flex flex-col items-center space-y-2 ",
-                router.pathname === "/live"
-                  ? "text-orange-500"
-                  : "hover:text-gray-500 transition-colors"
+                'flex flex-col items-center space-y-2 ',
+                router.pathname === '/live' ? 'text-orange-500' : 'hover:text-gray-500 transition-colors',
               )}
             >
               <svg
@@ -158,10 +140,8 @@ export default function Layout({
           <Link href="/profile">
             <a
               className={cls(
-                "flex flex-col items-center space-y-2 ",
-                router.pathname === "/profile"
-                  ? "text-orange-500"
-                  : "hover:text-gray-500 transition-colors"
+                'flex flex-col items-center space-y-2 ',
+                router.pathname === '/profile' ? 'text-orange-500' : 'hover:text-gray-500 transition-colors',
               )}
             >
               <svg
