+++ pages/enter.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import Button from '../components/button';
 import Input from '../components/input';
-import { cls } from '../libs/utils';
+import { cls } from '../libs/client/utils';
 
 interface EnterForm {
   email?: string;
